Different behaviors with the same presence condition are now differentiated.
The issue was caused by optional fragments, because the transformation
code was considering the fragment's guard as RDG node id.
Now the attribute being considered is the fragment's name or (in case it
is absent in the model) the fragment's id.

Note that this, however, immediately breaks the reuse we had up until
now, because we can no longer assume RDGNodes to be reusable whenever
they share the presence condition.
This issue will be subject to further performance tuning.

In order to be able to humanly identify RDG nodes, a toString() method
was added to the RDGNode class which combines the node's id with its
presence condition.