The infrastructure for representing SPL's behavioral models are ready.
The classes for creating Activity and Sequence Diagrams for representing the behavior of
software product lines are ready. The behavioral models are also being persisted into an
XML file for each software product line generated.

However it's necessary a major refactoring at the classes, in order to dispose the elements
in a better fashion. It's also necessary to ensure the tests for persistence works accordingly
and ensure XML contains only the elements for the software product line created (currently, it
represents the behavioral elements of other software product lines when two or more software
product lines are generated jointly. The problem is consequence of using the static elements
instead of the objects elements).

Next steps:
- refactor the classes of behavioral models generator
- ensure the XML representation contains only the elements of the generated software product line
- create a 'randomizer' classes which takes the input parameters and create the behavioral model
- represent cross-tree constraints accordingly
- format XML file of feature model in an affordable way
- create a graphical representation of the behavioral model
- integrate with reana, for performing the automated reliability analysis

 Changes to be committed:
	modified:   Activity.java
	modified:   ActivityDiagram.java
	modified:   ActivityDiagramElement.java
	modified:   DecisionNode.java
	modified:   EndNode.java
	modified:   FeatureModel.java
	modified:   Fragment.java
	modified:   Lifeline.java
	modified:   MergeNode.java
	modified:   Message.java
	new file:   SPL.java
	modified:   SequenceDiagram.java
	modified:   SequenceDiagramElement.java
	modified:   StartNode.java
	modified:   Transition.java